Implement camera system that tracks player; scroll screen using modulo calculation to wrap tilemap as needed in all directions

+++ sample-game/src/entities/directional_sprite.rs
@@ -58,7 +58,7 @@ impl DirectionalSprite {
         }
     }
 
-    pub fn draw(&self, renderer: &mut Renderer, x: usize, y: usize) {
+    pub fn draw(&self, renderer: &mut Renderer, x: i32, y: i32) {
         let tile = self
             .tileset
             .tiles
